amend tests to environment vars & selenium path
