+++ .env
@@ -1,0 +1,3 @@
+VUE_APP_API_ENDPOINT={Endpoint}
+ROOT=http://localhost:8080/
+TITLE=rekommend-manage
